requirements.txt
aliyun-python-sdk-core-v3==2.11.0
aliyun-python-sdk-dysmsapi==1.0.0
amqp==1.4.9
ansicolors==1.1.8
anyjson==0.3.3
appdirs==1.4.3
astroid==1.6.2
attrs==18.1.0
Babel==2.6.0
backcall==0.1.0
billiard==3.3.0.23
black==18.5b0
borax==1.2.0
celery==3.1.26.post2
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.7.3
construct==2.5.3
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
cycler==0.10.0
decorator==4.3.0
diff-match-patch==20121119
Django==2.2
django-celery==3.2.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-echarts==0.3.6
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-gulp==4.1.0
django-haystack==2.8.1
django-import-export==1.0.0
django-js-asset==1.2.2
django-memcached==0.1.2
django-mptt==0.9.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-rest-swagger==2.2.0
django-reversion==2.0.13
django-shortuuidfield==0.1.3
django-simple-captcha==0.5.6
djangorestframework==3.9.0
drf-haystack==1.8.4
dukpy==0.2.2
env-tools==2.2.0
et-xmlfile==1.0.1
flake8==3.6.0
Flask==1.0.2
Flask-Login==0.4.1
flower==0.9.2
freeze==1.0.10
future==0.16.0
httplib2==0.9.2
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
itypes==1.1.0
javascripthon==0.10
jdcal==1.3
jedi==0.12.0
jieba==0.39
Jinja2==2.10
jupyter-echarts-pypkg==0.1.2
kiwisolver==1.0.1
kombu==3.0.37
lazy-object-proxy==1.3.1
lml==0.0.2
macropy3==1.1.0b2
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.3
mccabe==0.6.1
mysqlclient==1.3.13
numpy==1.16.3
odfpy==1.3.6
openapi-codec==1.3.2
opencv-python==4.1.0.25
openpyxl==2.5.1
parso==0.2.1
pefile==2018.8.8
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.0.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
psutil==5.6.0
ptyprocess==0.6.0
pycodestyle==2.4.0
pycryptodome==3.7.0
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyecharts-snapshot==0.1.10
pyflakes==2.0.0
Pygments==2.2.0
pylint==1.8.3
PyMySQL==0.8.1
pyparsing==2.4.0
python-dateutil==2.8.0
python-memcached==1.59
python-ptrace==0.9.3
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
shortuuid==0.5.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
sqlparse==0.2.4
state==0.1.1.dev0
style==1.1.0
tablib==0.12.1
tini==3.0.1
tornado==5.1.1
traitlets==4.3.2
unicodecsv==0.14.1
update==0.0.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.14.1
Whoosh==2.7.4
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
yapf==0.25.0
